refactor: 파일명과 객체명 일치 및 안 쓰이는 부분 정리
