+++ src/components/icon24/AddFriendInac24.js
