+++ src/components/icon24/HeartInac24.js
