+++ src/pages/HomePages/EventStyles.js
