+++ src/pages/HomePages/HomePage.js
@@ -1,27 +1,26 @@
 import React, { useState } from 'react';
 import { LinearGradient } from 'expo-linear-gradient';
 import { View, Text, SafeAreaView, TouchableOpacity } from 'react-native';
-
-import HomeStyles from './homeStyles.js';
-
 import { useNavigation } from '@react-navigation/native';
 
-import HomeBg from '../../assets/images/svg_js/home_bg.js';
-import LogoBr from '../../components/logo/logo_br.js';
-import Notify from '../../components/icon_32/notification_32.js';
-import Homecard from '../../components/homeCompo/Homecard.js';
-import HomecardDifeF from '../../components/homeCompo/homecard_dife_f.js';
-import HomeSchEv from '../../components/homeCompo/HomeSchEv.js';
-import HomeProfile from '../../components/homeCompo/HomeProfile.js';
-import HomeSchoolInfo from '../../components/homeCompo/HomeScoolInfo.js';
-import HomeEvent from '../../components/homeCompo/HomeEvent.js';
-import HomeTag from '../../components/homeCompo/homeTag.js';
-import HomeArrow from '../../components/homeCompo/HomeArrow.js';
-import HeartInac24 from '../../components/icon_24/heart_inac_24.js';
-import AddFriendInac24 from '../../components/icon_24/addFriend_inac_24.js';
-import ChatInac24 from '../../components/icon_24/chat_inac_24.js';
-import HomeLine from '../../components/homeCompo/HomeLine.js';
-import HomecardBackBtn from '../../components/homeCompo/HomecardBackBtn.js';
+import HomeStyles from './HomeStyles.js';
+
+import HomeBg from '../../assets/images/svg_js/HomeBg.js';
+import LogoBr from '../../components/logo/LogoBr.js';
+import Notification32 from '../../components/icon32/Notification32.js';
+import Homecard from '../../components/HomeCompo/Homecard.js';
+import HomecardDifeF from '../../components/HomeCompo/HomecardDifeF.js';
+import HomeSchEv from '../../components/HomeCompo/HomeSchEv.js';
+import HomeProfile from '../../components/HomeCompo/HomeProfile.js';
+import HomeSchoolInfo from '../../components/HomeCompo/HomeScoolInfo.js';
+import HomeEvent from '../../components/HomeCompo/HomeEvent.js';
+import HomeTag from '../../components/HomeCompo/HomeTag.js';
+import HomeArrow from '../../components/HomeCompo/HomeArrow.js';
+import HeartInac24 from '../../components/icon24/HeartInac24.js';
+import AddFriendInac24 from '../../components/icon24/AddFriendInac24.js';
+import ChatInac24 from '../../components/icon24/ChatInac24.js';
+import HomeLine from '../../components/HomeCompo/HomeLine.js';
+import HomecardBackBtn from '../../components/HomeCompo/HomecardBackBtn.js';
 
 const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidunt ex at ullamcorper cursus odio laoreet placerat.",
     name = "Amyyheart",
@@ -39,14 +38,14 @@ const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidun
   return (
     <SafeAreaView style={HomeStyles.container}>
       <LinearGradient style={HomeStyles.linearGradient} colors={['#0029F4', '#6199C1', '#6199C1']}>
-        <HomeBg style={HomeStyles.svg} preserveAspectRatio='none'/>
+        <HomeBg style={HomeStyles.homebg} preserveAspectRatio='none' />
         
         <View style={HomeStyles.topContainer}>
           <View style={HomeStyles.logo}>
             <LogoBr />
           </View>
           <TouchableOpacity style={HomeStyles.notify} onPress={() => navigation.navigate('Notification')}>
-            <Notify />
+            <Notification32 />
           </TouchableOpacity>
         </View>
 
@@ -98,7 +97,7 @@ const HomePage = ({introduion = "adipiscing varius eu sit nulla, luctus tincidun
               </View>
               <Text style={HomeStyles.introduction}>{introduion}</Text>
               <View style={HomeStyles.myinfoContainer}>
-                <Text style={[HomeStyles.TXname, { fontFamily: 'NotoSansCJKkr-Medium' }]}>{name}</Text>
+                <Text style={HomeStyles.TXname}>{name}</Text>
                 <Text style={HomeStyles.myinfo}> | {country} | {age}</Text>
               </View>
               </View>
