+++ src/pages/HomePages/NotificationStyles.js
